requirements.txt
DateTime==4.3
easydict==1.9
imageio==2.6.1
matplotlib==3.0.3
numpy==1.16.2
opencv-contrib-python==3.4.2.16
progressbar==2.5
pyparsing==2.3.1
pypng==0.0.20
PyYAML==5.1
scikit-image==0.14.2
tensorboardX==1.9
torch==1.1.0
torchvision==0.2.0
tqdm==4.31.1